Add files via upload
In my latest book, Artificial Intelligence and Barbarism: A Critique of Digital Reason, I argue for a critical approach to AI, Technoskepticism based on Digital Humanism.
In February, the DISCO Network released its book “Technoskepticism, Between Possibility and Refusal” as a survival strategy for marginalized groups against systemic inequity.
The consensus is clear: we must be skeptical of Artificial Intelligence and see technology as inherently linked to power. However, the differences are also important for a more political interpretation of democratic technoskepticism.